refactor merge class
